Now pytest gets --island 1.2.3.4:5000 as command line param
