--- envs/monkey_zoo/blackbox/config.py
@@ -1,2 +1,0 @@
-ISLAND_SERVER_ADDRESS = "1.2.3.4"
-ISLAND_SERVER_URL = "https://{IP}/".format(IP=ISLAND_SERVER_ADDRESS)
